backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
boto3>=1.34.0
Pillow>=10.3.0
python-dotenv>=1.0.0
pytest>=8.2.0
